src/test/java/leetcode/lc0200_numberofislands/SolutionTest.java
package leetcode.lc0200_numberofislands;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

class SolutionTest {
    @Test
    void example1() {
        char[][] input = {
                {'1', '1', '1', '1', '0'},
                {'1', '1', '0', '1', '0'},
                {'1', '1', '0', '0', '0'},
                {'0', '0', '0', '0', '0'}
        };
        int expected = 1;

        int output = new Solution().numIslands(input);
        assertEquals(expected, output);
    }

    @Test
    void example2() {
        char[][] input = {
                {'1', '1', '0', '0', '0'},
                {'1', '1', '0', '0', '0'},
                {'0', '0', '1', '0', '0'},
                {'0', '0', '0', '1', '1'}
        };
        int expected = 3;

        int output = new Solution().numIslands(input);
        assertEquals(expected, output);
    }
}